AI-NEWSLETTER-GENERATOR/requirements.txt
streamlit
langchain
langchain-groq
langchain-core
langchain-community
sentence-transformers
faiss-cpu
google-serper
unstructured
pypdf
python-dotenv
pandas
openai
